pyproject.toml
[project]
name = "crispy-daisyui"
version = "0.13.0"
description = "A DaisyUI package for Django Crispy Forms"
readme = "README.md"
authors = [{name = "Fabian Geiger"}]
license = "Apache-2.0"
keywords = ["forms", "django", "crispy", "tailwind", "daisyui"]
requires-python = ">=3.7"
dependencies = [
    "django-crispy-forms>=1.11.2",
    "django>=3.2",
]

[project.urls]
Homepage = "https://github.com/fabge/crispy-daisyui"
Changelog = "https://github.com/fabge/crispy-daisyui/releases"
Issues = "https://github.com/fabge/crispy-daisyui/issues"
CI = "https://github.com/fabge/crispy-daisyui/actions"

[tool.setuptools.package-data]
crispy_daisyui = ["templates/**/*.html"]

[tool.uv]
package = true
